requirements.txt
streamlit
geopandas
matplotlib
streamlit-extras
# otras dependencias
